allow tables with tbody rows longer than thead
The way that GitHub.com does this (truncating all rows that are longer
than the thead row) is stupid. It hides content that could be important
to readers, without so much as a parser error. And that can be a pain
when you're not carefully checking what your markdown looks like after
compilation. For example, an extra cell jutting out of an otherwise
rectangular table, or a missing header is an easy error to spot, while a
column missing still leaves the table looking correct and orderly.

Even if you are looking at the output in a graphical/rendered form while
you're editing, it's annoying to start a table cell in a new column and
not be able to see what you are typing until you add a header for that
column.